requirements-dev.txt
pytest==8.2.2
requests==2.32.3
ruff==0.6.9
python-dotenv==1.2.2
httpx==0.28.1
typer==0.24.1
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
